Renamed Graphics Manager
Renamed graphics manager to comply with Pythonic naming convention of modules.
