+++ src/arcade/common/graphics_manager.py
